Rename `HTTPLinkParser` to `AtomURLResolver`
